centralize generic functions, add genericNew
